@@ -3,18 +3,23 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node_modules/react-native/packager/packager.sh",
+    "postinstall": "yarn run clean:example; yarn run clean:babelrc",
+    "clean:example": "rm -rf node_modules/react-native-gifted-chat/{node_modules,example}",
+    "clean:babelrc": "(find node_modules -type f -name .babelrc | grep -v /react-native/ | xargs rm) || true",
+    "start": "node_modules/react-native/scripts/launchPackager.command",
     "sync": "rm -rf ./node_modules/react-native-gifted-chat; sane '/usr/bin/rsync -v -a --exclude .git --exclude example --exclude __tests__ --exclude node_modules ../ ./node_modules/react-native-gifted-chat/' .. --glob='{**/*.json,**/*.js}'"
   },
   "dependencies": {
-    "react": "~15.4.0-rc.4",
-    "react-native": "^0.40.0",
-    "react-native-camera-roll-picker": "^1.1.7",
-    "react-native-gifted-chat": "file:../",
-    "react-native-nav": "^1.1.4"
+    "prop-types": "15.5.10",
+    "react": "16.0.0-alpha.12",
+    "react-native": "0.46.3",
+    "react-native-camera-roll-picker": "1.1.7",
+    "react-native-gifted-chat": "file:..",
+    "react-native-maps": "0.15.3",
+    "react-native-nav": "cribspot/react-native-nav#76baa81ded23829ad9b5eb4e4a51052b20fb4e08"
   },
   "devDependencies": {
-    "babel-preset-react-native": "^1.9.0",
-    "sane": "^1.4.0"
+    "babel-preset-react-native": "2.1.0",
+    "sane": "2.0.0"
   }
 }
