@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {
   Platform,
@@ -6,12 +7,13 @@ import {
 } from 'react-native';
 
 export default class Composer extends React.Component {
-  onChange(e) {
+  onContentSizeChange(e) {
     const contentSize = e.nativeEvent.contentSize;
-    if (!this.contentSize) {
-      this.contentSize = contentSize;
-      this.props.onInputSizeChanged(this.contentSize);
-    } else if (this.contentSize.width !== contentSize.width || this.contentSize.height !== contentSize.height) {
+
+    // Support earlier versions of React Native on Android.
+    if (!contentSize) return;
+
+    if (!this.contentSize || this.contentSize.width !== contentSize.width || this.contentSize.height !== contentSize.height) {
       this.contentSize = contentSize;
       this.props.onInputSizeChanged(this.contentSize);
     }
@@ -28,14 +30,18 @@ export default class Composer extends React.Component {
         placeholderTextColor={this.props.placeholderTextColor}
         multiline={this.props.multiline}
 
-        onChange={(e) => this.onChange(e)}
-        onChangeText={text => this.onChangeText(text)}
+        onChange={(e) => this.onContentSizeChange(e)}
+        onContentSizeChange={(e) => this.onContentSizeChange(e)}
 
-        style={[styles.textInput, this.props.textInputStyle, {height: this.props.composerHeight}]}
+        onChangeText={(text) => this.onChangeText(text)}
+
+        style={[styles.textInput, this.props.textInputStyle, { height: this.props.composerHeight }]}
+
+        autoFocus={this.props.textInputAutoFocus}
 
         value={this.props.text}
         accessibilityLabel={this.props.text || this.props.placeholder}
-        enablesReturnKeyAutomatically={true}
+        enablesReturnKeyAutomatically
         underlineColorAndroid="transparent"
         {...this.props.textInputProps}
       />
@@ -61,18 +67,16 @@ const styles = StyleSheet.create({
 });
 
 Composer.defaultProps = {
-  onChange: () => {
-  },
   composerHeight: Platform.select({
     ios: 33,
     android: 41,
   }), // TODO SHARE with GiftedChat.js and tests
   text: '',
-  placeholder: 'Type a message...',
   placeholderTextColor: '#b2b2b2',
   textInputProps: null,
   multiline: true,
   textInputStyle: {},
+  textInputAutoFocus: false,
   onTextChanged: () => {
   },
   onInputSizeChanged: () => {
@@ -80,14 +84,14 @@ Composer.defaultProps = {
 };
 
 Composer.propTypes = {
-  onChange: React.PropTypes.func,
-  composerHeight: React.PropTypes.number,
-  text: React.PropTypes.string,
-  placeholder: React.PropTypes.string,
-  placeholderTextColor: React.PropTypes.string,
-  textInputProps: React.PropTypes.object,
-  onTextChanged: React.PropTypes.func,
-  onInputSizeChanged: React.PropTypes.func,
-  multiline: React.PropTypes.bool,
+  composerHeight: PropTypes.number,
+  text: PropTypes.string,
+  placeholder: PropTypes.string,
+  placeholderTextColor: PropTypes.string,
+  textInputProps: PropTypes.object,
+  onTextChanged: PropTypes.func,
+  onInputSizeChanged: PropTypes.func,
+  multiline: PropTypes.bool,
   textInputStyle: TextInput.propTypes.style,
+  textInputAutoFocus: PropTypes.bool,
 };
